requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
streamlit-drawable-canvas==0.9.3
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.44.0
